requirements.txt
pydantic>=2.8
httpx>=0.27
tqdm>=4.66
rich>=13.7
python-dotenv>=1.0
